Move name validation to Pydantic
